+++ modules/tenants/src/main/java/com/eureckah/banking/tenants/application/port/out/messaging/TenantEventsPublisher.java
@@ -1,0 +1,8 @@
+package com.eureckah.banking.tenants.application.port.out.messaging;
+
+import com.eureckah.banking.tenants.domain.events.TenantEvent;
+
+public interface TenantEventsPublisher {
+
+    void publishTenantEvent(TenantEvent event);
+}