+++ modules/tenants/src/main/java/com/eureckah/banking/tenants/infrastructure/adapter/out/jpa/user/UserJpaEntity.java
@@ -1,0 +1,30 @@
+package com.eureckah.banking.tenants.infrastructure.adapter.out.jpa.user;
+
+import jakarta.persistence.*;
+
+import lombok.*;
+
+import org.hibernate.annotations.UuidGenerator;
+
+import java.util.UUID;
+
+@Getter
+@Entity
+@Table(name = "users")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder
+public class UserJpaEntity {
+
+    @Id
+    @GeneratedValue
+    @UuidGenerator
+    @Column(name = "id", nullable = false, updatable = false)
+    private UUID id;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @Column(name = "email", nullable = false, unique = true)
+    private String email;
+}